requirements.txt
boto3==1.35.41
requests==2.27.1
snowflake-connector-python==3.12.2
pyspark==3.5.3
urllib3==2.2.3
pytest==8.3.3